+++ src/test/java/be/woutschoovaerts/mollie/handler/MethodHandlerIntegrationTest.java
@@ -4,13 +4,17 @@ import be.woutschoovaerts.mollie.Client;
 import be.woutschoovaerts.mollie.ClientBuilder;
 import be.woutschoovaerts.mollie.data.common.Pagination;
 import be.woutschoovaerts.mollie.data.method.MethodListResponse;
+import be.woutschoovaerts.mollie.data.method.MethodPricing;
 import be.woutschoovaerts.mollie.data.method.MethodResponse;
 import be.woutschoovaerts.mollie.exception.MollieException;
+import be.woutschoovaerts.mollie.util.QueryParams;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.util.List;
+
 import static be.woutschoovaerts.mollie.IntegrationTestConstants.API_KEY;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.*;
 
 class MethodHandlerIntegrationTest {
 
@@ -40,5 +44,33 @@ class MethodHandlerIntegrationTest {
         MethodResponse method = client.methods().getMethod("ideal");
 
         assertNotNull(method);
+    }
+
+    @Test
+    void getMethodPricingWithFeeRegion() throws MollieException {
+        QueryParams queryParams = new QueryParams();
+        queryParams.put("include", "pricing");
+
+        MethodResponse method = client.methods().getMethod("creditcard", queryParams);
+        List<MethodPricing> pricing = method.getPricing();
+
+        assertNotEquals(0, pricing.size());
+        for (MethodPricing methodPricing: pricing) {
+            assertTrue(methodPricing.getFeeRegion().isPresent());
+        }
+    }
+
+    @Test
+    void getMethodPricingWithoutFeeRegion() throws MollieException {
+        QueryParams queryParams = new QueryParams();
+        queryParams.put("include", "pricing");
+
+        MethodResponse method = client.methods().getMethod("ideal", queryParams);
+        List<MethodPricing> pricing = method.getPricing();
+
+        assertNotEquals(0, pricing.size());
+        for (MethodPricing methodPricing: pricing) {
+            assertFalse(methodPricing.getFeeRegion().isPresent());
+        }
     }
 }